Implemented JWT and Permissions with Some Reorganization

+++ database/users.py
@@ -1,12 +1,14 @@
 # Imports
+from config.config import permissions
 from .base import DataAccessBase
-from models.user import User
 from utils.hash import sha256
+from models.user import User
 import uuid
 
 class UserAccess(DataAccessBase):
     """Class that handles user information"""
     
+    @staticmethod
     def get_user(secure=False, obj=False, **kwargs):
         """Base method for get_user methods"""
         
@@ -67,7 +69,7 @@ class UserAccess(DataAccessBase):
             return {"status": "success", "message": "User added to the system"}
         # Return false if the given information exists
         else:
-            return {"status": "error", "message": "User didn't register"}
+            return {"status": "error", "message": "User did not register"}
         
     @staticmethod
     def register_user(**kwargs):
@@ -109,7 +111,7 @@ class UserAccess(DataAccessBase):
             return check
 
         # Check if the permission value is a list 
-        if not isinstance(kwargs["permission"], list):
+        if not isinstance(kwargs["permissions"], list):
             return {"status": "error", "message": "Permission value not in list format"}
         
         # Check if the operation value is one of the accepted options
@@ -121,13 +123,21 @@ class UserAccess(DataAccessBase):
         
         # Add new permission(s) and track changes
         results = {}
-        for permission in kwargs["permission"]:
+        for permission in kwargs["permissions"]:
+            # If the given permission is not part of the approved list of permission
+            # track that is is not added with an exaplanation and continue
+            if permission not in permissions:
+                results[permission] = "Not Added (Invalid Permission)"
+                continue
+            
             # Attempt add permission
             res = user.add_permission(permission) if operation == "add" else \
                 user.delete_permission(permission)
             
             # Track changes
-            results[permission] = res
+            results[permission] = ("Added" if operation == "add" else "Deleted") \
+                if res else "Not Added (Already " + ("Added)" if operation == "add" \
+                else "Deleted or is Missing)")
         
         # Update database
         DataAccessBase.USER_COL.update_one(
